Cleanup table 'key' usage
